+++ src/felupe/__about__.py
@@ -1,1 +1,1 @@
-__version__ = "9.6.0-dev"
+__version__ = "10.0.0-dev"